Fix file moving to be safe on Windows.
Fix file moving to be safe on Windows. A shim was added
that calls into the Win32 API / MoveFileEx that prevents an error
occurring on windows if the destination already exists. This shim might
not be needed once os.Replace lands, possibly in Go 1.5.

See: https://github.com/golang/go/issues/8914

[#88907658]